requirements.txt
Flask>=2.2,<2.3
python-dotenv>=0.21,<1.1
requests>=2.28,<3
